requirements.txt
requests==2.9.1
pyyaml==3.11
